Added player movement, smooth camera follow and messed around with gravity
For my game I only require left and right player movement so that is all I decided to implement. I got the camera to follow the player smoothly as its important to the gameplay. I have looked into the player gravity as I want the player to fall slowly, as my game will be focused on falling and avoiding obstacles. I haven't gotten it just right yet.