Fixed bugs + added tilemaps
I decided to revert to older movement and gun scripts as I was having many issues with the previous scripts. I created the map which is now filled with objects. I changed some of the objects colliders, for example I edited the collision on trees so that the player wouldn't collide with the top half of a tree as it made it more realistic