Added Enemies
I created 2 types of enemies, one is slower than the other and does more damage. I watched a tutorial on how to make the enemies follow the player. I also created a EnemyData script which stores an enemies hp, damage and speed. this script can be applied to any enemy